requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
PyJWT==2.6.0
pymongo==3.12.3
pytz==2022.7.1
sqlparse==0.2.4
